refactor(e2e): unify fixture import paths for consistency
Updated fixture import paths in multiple test files to use a unified `./db` path instead of individual database files. This change enhances clarity and maintainability across the end-to-end testing framework.
